refact: use httpx instead of request
