+++ lybic/lmcp.py
@@ -25,13 +25,14 @@
 # THE SOFTWARE.
 
 """lmcp.py: MCP client for lybic MCP(Model Context Protocol) and Restful Interface API."""
-import asyncio
+from mcp import ClientSession
+from mcp.client.streamable_http import streamablehttp_client
 
 from lybic import dto
 from lybic.lybic import LybicClient
 
 class MCP:
-    """MCP is a client for lybic MCP(Model Context Protocol) and Restful Interface API."""
+    """AsyncMCP is a client for lybic MCP(Model Context Protocol) and Restful Interface API."""
     def __init__(self, client: LybicClient):
         """
         Init MCP client with lybic client
@@ -40,41 +41,51 @@ class MCP:
         """
         self.client = client
 
-    def list(self) -> dto.ListMcpServerResponse:
+    async def list(self) -> dto.ListMcpServerResponse:
         """
         List all MCP servers in the organization
 
         :return:
         """
-        return asyncio.run(self.client.get_async_client().mcp.list())
+        response = await self.client.request(
+            "GET",
+            f"/api/orgs/{self.client.org_id}/mcp-servers")
+        return dto.ListMcpServerResponse.model_validate_json(response.text)
 
-    def create(self, data: dto.CreateMcpServerDto) -> dto.McpServerResponseDto:
+    async def create(self, data: dto.CreateMcpServerDto) -> dto.McpServerResponseDto:
         """
         Create a mcp server
 
         :param data:
         :return:
         """
-        return asyncio.run(self.client.get_async_client().mcp.create(data))
+        response = await self.client.request(
+            "POST",
+            f"/api/orgs/{self.client.org_id}/mcp-servers",
+            json=data.model_dump())
+        return dto.McpServerResponseDto.model_validate_json(response.text)
 
-    def get_default(self) -> dto.McpServerResponseDto:
+    async def get_default(self) -> dto.McpServerResponseDto:
         """
         Get default mcp server
 
         :return:
         """
-        return asyncio.run(self.client.get_async_client().mcp.get_default())
+        response = await self.client.request(
+            "GET",
+            f"/api/orgs/{self.client.org_id}/mcp-servers/default")
+        return dto.McpServerResponseDto.model_validate_json(response.text)
 
-    def delete(self, mcp_server_id: str) -> None:
+    async def delete(self, mcp_server_id: str) -> None:
         """
         Delete a mcp server
 
         :param mcp_server_id:
         :return:
         """
-        return asyncio.run(self.client.get_async_client().mcp.delete(mcp_server_id))
+        await self.client.request("DELETE", f"/api/orgs/{self.client.org_id}/mcp-servers/{mcp_server_id}")
 
-    def set_sandbox(self, mcp_server_id: str, sandbox_id: str) -> None:
+    async def set_sandbox(self, mcp_server_id: str, sandbox_id: str) -> None:
         """
         Set MCP server to a specific sandbox
 
@@ -83,7 +94,10 @@ class MCP:
         :return: None
         """
         data = dto.SetMcpServerToSandboxResponseDto(sandboxId=sandbox_id)
-        return asyncio.run(self.client.get_async_client().mcp.set_sandbox(mcp_server_id, data))
+        await self.client.request(
+            "POST",
+            f"/api/orgs/{self.client.org_id}/mcp-servers/{mcp_server_id}/sandbox",
+            json=data.model_dump())
 
     async def call_tool_async(self,
                               mcp_server_id: str,
@@ -97,27 +111,47 @@ class MCP:
         :param tool_args:
         :return:
         """
-        return self.client.get_async_client().mcp.call_tool_async(mcp_server_id, tool_name, tool_args)
+        try:
+            async with streamablehttp_client(self.client.make_mcp_endpoint(mcp_server_id),
+                                             headers=self.client.headers,
+                                             timeout=30
+            ) as (
+                    read_stream,
+                    write_stream,
+                    _,
+            ):
+                async with ClientSession(read_stream, write_stream) as session:
+                    await session.initialize()
+                    return await session.call_tool(tool_name, tool_args)
+        except Exception as e:
+            raise RuntimeError(f"Failed to call tool: {e}") from e
 
 class ComputerUse:
-    """ComputerUse is a client for lybic ComputerUse API(MCP and Restful)."""
+    """AsyncComputerUse is a client for lybic ComputerUse API(MCP and Restful)."""
     def __init__(self, client: LybicClient):
         self.client = client
 
-    def parse_model_output(self, data: dto.ComputerUseParseRequestDto) -> dto.ComputerUseActionResponseDto:
+    async def parse_model_output(self, data: dto.ComputerUseParseRequestDto) -> dto.ComputerUseActionResponseDto:
         """
         parse doubao-ui-tars output
 
         :param data:
         :return:
         """
-        return asyncio.run(self.client.get_async_client().computer_use.parse_model_output(data))
+        response = await self.client.request(
+            "POST",
+            "/api/computer-use/parse",
+            json=data.model_dump(exclude_none=True))
+        return dto.ComputerUseActionResponseDto.model_validate_json(response.text)
 
-    def execute_computer_use_action(self, sandbox_id: str,
+    async def execute_computer_use_action(self, sandbox_id: str,
                                     data: dto.ComputerUseActionDto) -> dto.SandboxActionResponseDto:
         """
         Execute a computer use action
 
         is same as sandbox.Sandbox.execute_computer_use_action
         """
-        return asyncio.run(self.client.get_async_client().computer_use.execute_computer_use_action(sandbox_id, data))
+        response = await self.client.request("POST",
+                                       f"/api/orgs/{self.client.org_id}/sandboxes/{sandbox_id}/actions/computer-use",
+                                       json=data.model_dump(exclude_none=True))
+        return dto.SandboxActionResponseDto.model_validate_json(response.text)
