+++ requirements.txt
@@ -1,6 +1,5 @@
-requests>=2.32.3
 pydantic>=2.10.6
 mcp>=1.12.0
 pillow==11.3.0
+httpx==0.28.1
 setuptools
-httpx
